fixed sorting and logs show clashing error, app functions as should
