+++ src/Components/Provider.js
@@ -1,4 +1,5 @@
-import { createContext, useState} from "react";
+import { createContext, useState, useEffect} from "react";
+import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import Nav from "./Nav";
 import Footer from "./Footer";
@@ -10,6 +11,17 @@ function Provider({children}) {
     const API = process.env.REACT_APP_API_URL
     const [deleteModal, setDeleteModal] = useState(false)
     const [modalIndex, setModalIndex] = useState("")
+    // move logs state here due to sort dropdown interference
+    const [logs, setLogs] = useState([])
+    // move select state here so can check value in logsShow component
+    const [select, setSelect] = useState("default")
+    const navigate = useNavigate()
+
+    useEffect(() => {
+        axios.get(`${API}`)
+        .then(respJson => setLogs(respJson.data))
+        .catch(err => navigate("/*"))
+    },[])
 
     return (
             <ContextData.Provider value={{
@@ -18,7 +30,11 @@ function Provider({children}) {
                 deleteModal,
                 setDeleteModal,
                 modalIndex,
-                setModalIndex, 
+                setModalIndex,
+                logs, 
+                setLogs,
+                select,
+                setSelect, 
             }}>
                 <Nav />
                 <Footer />
